requirements.txt
openai
python-telegram-bot
pillow
python-dotenv
ruff